code cleaned up

+++ src/Channels.cpp
@@ -17,7 +17,7 @@ namespace dehancer {
     
     size_t ChannelsDesc::get_hash () const {
       size_t cx = 0;
-      for (int i = 0; i < scale.size(); ++i) {
+      for (size_t i = 0; i < scale.size(); ++i) {
         cx += static_cast<size_t>( scale[i].x * 10000 + scale[i].y * 100 ) << i;
       }
       return
@@ -42,15 +42,6 @@ namespace dehancer {
             std::shared_ptr<ChannelItem> item_;
             ChannelsDesc desc_;
             std::shared_ptr<std::array<ChannelsDesc,4>> channel_descs_;
-//
-//            void set_active_mask(const ChannelsHolder::ActiveChannelsMask& amask) override {
-//              //item_->amask = amask;
-//              //init();
-//            };
-//
-//            [[nodiscard]] const ChannelsHolder::ActiveChannelsMask& get_active_mask() const override {
-//              return item_->amask;
-//            };
             
             size_t get_width(int index) const override { return channel_descs_->at(index).width; };
             size_t get_height(int index) const override {return channel_descs_->at(index).height;};
@@ -284,7 +275,7 @@ namespace dehancer {
     
     void ChannelsInput::set_scale (ChannelsDesc::Scale2D scale) {
       bool do_recreate_channels = false;
-      for (int i = 0; i < scale.size(); ++i) {
+      for (size_t i = 0; i < scale.size(); ++i) {
         if (impl_->desc.scale.at(i).x!=scale.at(i).y || impl_->desc.scale.at(i).y!=scale.at(i).y) {
           do_recreate_channels = true;
           break;
@@ -367,11 +358,7 @@ namespace dehancer {
         
         auto channel = channels->at(j);
         
-        //bool has_channel = channel != nullptr;
-        
         if (!channel) continue;
-  
-        //channel = MemoryHolder::Make(get_command_queue(),1);
         
         execute([this, channels, &channel, j](CommandEncoder& encoder){
